vk/gl: Factorize shared GLSL code - prep vulkan for shared glsl backend
