+++ rpcs3/Emu/RSX/VK/VKCommonDecompiler.cpp
@@ -8,23 +8,6 @@ namespace vk
 {
 	static TBuiltInResource g_default_config;
 
-	std::string getFloatTypeNameImpl(size_t elementCount)
-	{
-		switch (elementCount)
-		{
-		default:
-			abort();
-		case 1:
-			return "float";
-		case 2:
-			return "vec2";
-		case 3:
-			return "vec3";
-		case 4:
-			return "vec4";
-		}
-	}
-
 	std::string getFunctionImpl(FUNCTION f)
 	{
 		switch (f)
@@ -82,66 +65,6 @@ namespace vk
 		case FUNCTION::FUNCTION_TEXTURE_SAMPLE2D_DEPTH_RGBA:
 			return "texture2DReconstruct($t, $0.xy)";
 		}
-	}
-
-	std::string compareFunctionImpl(COMPARE f, const std::string &Op0, const std::string &Op1)
-	{
-		switch (f)
-		{
-		case COMPARE::FUNCTION_SEQ:
-			return "equal(" + Op0 + ", " + Op1 + ")";
-		case COMPARE::FUNCTION_SGE:
-			return "greaterThanEqual(" + Op0 + ", " + Op1 + ")";
-		case COMPARE::FUNCTION_SGT:
-			return "greaterThan(" + Op0 + ", " + Op1 + ")";
-		case COMPARE::FUNCTION_SLE:
-			return "lessThanEqual(" + Op0 + ", " + Op1 + ")";
-		case COMPARE::FUNCTION_SLT:
-			return "lessThan(" + Op0 + ", " + Op1 + ")";
-		case COMPARE::FUNCTION_SNE:
-			return "notEqual(" + Op0 + ", " + Op1 + ")";
-		}
-		fmt::throw_exception("Unknown compare function" HERE);
-	}
-
-	void insert_glsl_legacy_function(std::ostream& OS, glsl::program_domain domain)
-	{
-		OS << "vec4 lit_legacy(vec4 val)";
-		OS << "{\n";
-		OS << "	vec4 clamped_val = val;\n";
-		OS << "	clamped_val.x = max(val.x, 0.);\n";
-		OS << "	clamped_val.y = max(val.y, 0.);\n";
-		OS << "	vec4 result;\n";
-		OS << "	result.x = 1.;\n";
-		OS << "	result.w = 1.;\n";
-		OS << "	result.y = clamped_val.x;\n";
-		OS << "	result.z = clamped_val.x > 0. ? exp(clamped_val.w * log(max(clamped_val.y, 1.E-10))) : 0.;\n";
-		OS << "	return result;\n";
-		OS << "}\n\n";
-
-		if (domain == glsl::program_domain::glsl_vertex_program)
-			return;
-
-		//NOTE: After testing with GOW, the w component is either the original depth or wraps around to the x component
-		//Since component.r == depth_value with some precision loss, just use the precise depth value for now (further testing needed)
-		OS << "vec4 decodeLinearDepth(float depth_value)\n";
-		OS << "{\n";
-		OS << "	uint value = uint(depth_value * 16777215);\n";
-		OS << "	uint b = (value & 0xff);\n";
-		OS << "	uint g = (value >> 8) & 0xff;\n";
-		OS << "	uint r = (value >> 16) & 0xff;\n";
-		OS << "	return vec4(float(r)/255., float(g)/255., float(b)/255., depth_value);\n";
-		OS << "}\n\n";
-
-		OS << "vec4 texture2DReconstruct(sampler2D tex, vec2 coord)\n";
-		OS << "{\n";
-		OS << "	return decodeLinearDepth(texture(tex, coord.xy).r);\n";
-		OS << "}\n\n";
-
-		OS << "vec4 texture2DReconstruct(sampler2DRect tex, vec2 coord)\n";
-		OS << "{\n";
-		OS << "	return decodeLinearDepth(texture(tex, coord.xy).r);\n";
-		OS << "}\n\n";
 	}
 
 	void init_default_resources(TBuiltInResource &rsc)
@@ -274,9 +197,9 @@ namespace vk
 		fmt::throw_exception("Unknown register name: %s" HERE, name);
 	}
 
-	bool compile_glsl_to_spv(std::string& shader, glsl::program_domain domain, std::vector<u32>& spv)
+	bool compile_glsl_to_spv(std::string& shader, program_domain domain, std::vector<u32>& spv)
 	{
-		EShLanguage lang = (domain == glsl::glsl_fragment_program) ? EShLangFragment : EShLangVertex;
+		EShLanguage lang = (domain == glsl_fragment_program) ? EShLangFragment : EShLangVertex;
 
 		glslang::TProgram program;
 		glslang::TShader shader_object(lang);
